refactor: Move crypto config to config directory
