+++ apps/web/src/components/SearchModal/CurrencyList.tsx
@@ -11,7 +11,7 @@ import { FixedSizeList } from 'react-window'
 import { styled } from 'styled-components'
 import { wrappedCurrency } from 'utils/wrappedCurrency'
 import { useAccount } from 'wagmi'
-import { fiatCurrencyMap } from 'views/BuyCrypto/constants'
+import { fiatCurrencyMap } from 'config/constants/buyCrypto'
 import { FiatLogo } from 'components/Logo/CurrencyLogo'
 import { useIsUserAddedToken } from '../../hooks/Tokens'
 import { useCombinedActiveList } from '../../state/lists/hooks'
