+++ apps/web/src/components/SearchModal/CurrencySearch.tsx
@@ -11,7 +11,7 @@ import { WrappedTokenInfo, createFilterToken } from '@pancakeswap/token-lists'
 import { useAudioPlay } from '@pancakeswap/utils/user'
 import { isAddress } from 'utils'
 import { useActiveChainId } from 'hooks/useActiveChainId'
-import { whiteListedFiatCurrencies } from 'views/BuyCrypto/constants'
+import { whiteListedFiatCurrencies } from 'config/constants/buyCrypto'
 import { useAllTokens, useIsUserAddedToken, useToken } from '../../hooks/Tokens'
 import Row from '../Layout/Row'
 import CommonBases from './CommonBases'
