+++ apps/web/src/hooks/useCanBuyCrypto.ts
@@ -1,8 +1,6 @@
 import { ChainId } from '@pancakeswap/sdk'
 import { useMemo } from 'react'
-
-// TODO: refactor. We shouldn't import modules from views in higher level hooks
-import { SUPPORTED_ONRAMP_TOKENS, isBuyCryptoSupported } from 'views/BuyCrypto/constants'
+import { SUPPORTED_ONRAMP_TOKENS, isBuyCryptoSupported } from 'config/constants/buyCrypto'
 
 interface Params {
   currencySymbol?: string
