+++ apps/web/src/views/BuyCrypto/components/AccordionDropdown/AccordionItem.tsx
@@ -8,7 +8,7 @@ import styled from 'styled-components'
 import { useTranslation } from '@pancakeswap/localization'
 import { isMobile } from 'react-device-detect'
 import formatLocaleNumber from 'utils/formatLocaleNumber'
-import { providerFeeTypes } from 'views/BuyCrypto/constants'
+import { providerFeeTypes } from 'config/constants/buyCrypto'
 import Image from 'next/image'
 import getTimePeriods from '@pancakeswap/utils/getTimePeriods'
 import OnRampProviderLogo from '../OnRampProviderLogo/OnRampProviderLogo'
